// old/data/type.ts
export interface LOCATION_AND_DIMENSION {
	/**
	 * TODO: 位置对象
	 */
	location: server.Vector3,
	/**
	 * TODO: 维度对象
	 */
	dimension: server.Dimension
};